+++ lib/lightgbm/dataset.rb
@@ -2,7 +2,7 @@ module LightGBM
   class Dataset
     attr_reader :data, :params
 
-    def initialize(data, label: nil, weight: nil, group: nil, params: nil, reference: nil, used_indices: nil, categorical_feature: "auto", feature_names: nil)
+    def initialize(data, label: nil, weight: nil, group: nil, params: nil, reference: nil, used_indices: nil, categorical_feature: "auto", feature_name: nil, feature_names: nil)
       @data = data
       @label = label
       @weight = weight
@@ -11,7 +11,7 @@ module LightGBM
       @reference = reference
       @used_indices = used_indices
       @categorical_feature = categorical_feature
-      @feature_names = feature_names
+      @feature_name = feature_name || feature_names || "auto"
 
       construct
     end
@@ -24,7 +24,7 @@ module LightGBM
       field("weight")
     end
 
-    def feature_names
+    def feature_name
       # must preallocate space
       num_feature_names = ::FFI::MemoryPointer.new(:int)
       out_buffer_len = ::FFI::MemoryPointer.new(:size_t)
@@ -48,6 +48,7 @@ module LightGBM
       # from most recent call (instead of num_features)
       str_ptrs[0, num_feature_names.read_int].map(&:read_string)
     end
+    alias_method :feature_names, :feature_name
 
     def label=(label)
       @label = label
@@ -64,12 +65,15 @@ module LightGBM
       set_field("group", group, type: :int32)
     end
 
-    def feature_names=(feature_names)
+    def feature_name=(feature_names)
       @feature_names = feature_names
       c_feature_names = ::FFI::MemoryPointer.new(:pointer, feature_names.size)
-      c_feature_names.write_array_of_pointer(feature_names.map { |v| ::FFI::MemoryPointer.from_string(v) })
+      # keep reference to string pointers
+      str_ptrs = feature_names.map { |v| ::FFI::MemoryPointer.from_string(v) }
+      c_feature_names.write_array_of_pointer(str_ptrs)
       check_result FFI.LGBM_DatasetSetFeatureNames(handle_pointer, c_feature_names, feature_names.size)
     end
+    alias_method :feature_names=, :feature_name=
 
     # TODO only update reference if not in chain
     def reference=(reference)
@@ -106,12 +110,7 @@ module LightGBM
     end
 
     def handle_pointer
-      @handle.read_pointer
-    end
-
-    def self.finalize(addr)
-      # must use proc instead of stabby lambda
-      proc { FFI.LGBM_DatasetFree(::FFI::Pointer.new(:pointer, addr)) }
+      @handle
     end
 
     private
@@ -127,25 +126,33 @@ module LightGBM
       end
       set_verbosity(params)
 
-      @handle = ::FFI::MemoryPointer.new(:pointer)
+      handle = ::FFI::MemoryPointer.new(:pointer)
       parameters = params_str(params)
       reference = @reference.handle_pointer if @reference
       if used_indices
         used_row_indices = ::FFI::MemoryPointer.new(:int32, used_indices.count)
         used_row_indices.write_array_of_int32(used_indices)
-        check_result FFI.LGBM_DatasetGetSubset(reference, used_row_indices, used_indices.count, parameters, @handle)
+        check_result FFI.LGBM_DatasetGetSubset(reference, used_row_indices, used_indices.count, parameters, handle)
       elsif data.is_a?(String)
-        check_result FFI.LGBM_DatasetCreateFromFile(data, parameters, reference, @handle)
+        check_result FFI.LGBM_DatasetCreateFromFile(data, parameters, reference, handle)
       else
         if matrix?(data)
           nrow = data.row_count
           ncol = data.column_count
           flat_data = data.to_a.flatten
         elsif daru?(data)
+          if @feature_name == "auto"
+            @feature_name = data.vectors.to_a
+          end
           nrow, ncol = data.shape
           flat_data = data.map_rows(&:to_a).flatten
-        elsif numo?(data) || rover?(data)
-          data = data.to_numo if rover?(data)
+        elsif numo?(data)
+          nrow, ncol = data.shape
+        elsif rover?(data)
+          if @feature_name == "auto"
+            @feature_name = data.keys
+          end
+          data = data.to_numo
           nrow, ncol = data.shape
         else
           nrow = data.count
@@ -161,14 +168,18 @@ module LightGBM
           c_data.write_array_of_double(flat_data)
         end
 
-        check_result FFI.LGBM_DatasetCreateFromMat(c_data, 1, nrow, ncol, 1, parameters, reference, @handle)
+        check_result FFI.LGBM_DatasetCreateFromMat(c_data, 1, nrow, ncol, 1, parameters, reference, handle)
       end
-      ObjectSpace.define_finalizer(@handle, self.class.finalize(handle_pointer.to_i)) unless used_indices
+      if used_indices
+        @handle = handle.read_pointer
+      else
+        @handle = ::FFI::AutoPointer.new(handle.read_pointer, FFI.method(:LGBM_DatasetFree))
+      end
 
       self.label = @label if @label
       self.weight = @weight if @weight
       self.group = @group if @group
-      self.feature_names = @feature_names if @feature_names
+      self.feature_name = @feature_name if @feature_name && @feature_name != "auto"
     end
 
     def dump_text(filename)