+++ test/train_test.rb
@@ -54,7 +54,11 @@ class TrainTest < Minitest::Test
       model = LightGBM.train(regression_params, regression_train, valid_sets: [regression_train, regression_test], early_stopping_rounds: 500)
     end
     assert_equal 100, model.best_iteration
-    assert_includes stdout, "Best iteration is: [100]\ttraining's l2: 0.024524\tvalid_1's l2: 0.0841232"
+    if jruby?
+      assert_includes stdout, "Best iteration is: [100]\ttraining's l2: 0.0245240\tvalid_1's l2: 0.0841232"
+    else
+      assert_includes stdout, "Best iteration is: [100]\ttraining's l2: 0.024524\tvalid_1's l2: 0.0841232"
+    end
   end
 
   def test_early_stopping_early_higher_better